formChallenge/requirements.txt
dataclasses==0.6
Django==2.2.5
djangorestframework==3.10.3
djongo==1.2.35
dnspython==1.16.0
mongoengine==0.18.2
pymongo==3.9.0
pytz==2019.2
six==1.12.0
sqlparse==0.2.4
